requirements.txt
PySide6
mss
Pillow
pyautogui
numpy
opencv-python
